% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\docType{data}
\name{ERA5_day_sfcwind}
\alias{ERA5_day_sfcwind}
\title{ERA5 daily surface wind speed}
\format{
A climate4R grid object
}
\source{
ERA5 reanalysis data
}
\description{
A climate4R grid with daily surface wind speed (sfcwind) from ERA5.
}
\keyword{data}
