% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\docType{data}
\name{ERA5_day_ssrd}
\alias{ERA5_day_ssrd}
\title{ERA5 daily surface solar radiation downwards}
\format{
A climate4R grid object
}
\source{
ERA5 reanalysis data
}
\description{
A climate4R grid with daily surface solar radiation (ssrd) from ERA5.
}
\keyword{data}
