% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\docType{data}
\name{ERA5_day_t2mx}
\alias{ERA5_day_t2mx}
\title{ERA5 daily maximum 2m temperature}
\format{
A climate4R grid object
}
\source{
ERA5 reanalysis data
}
\description{
A climate4R grid with daily maximum 2m temperature (t2mx) from ERA5.
}
\keyword{data}
