% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\docType{data}
\name{ERA5_lsm}
\alias{ERA5_lsm}
\title{ERA5 land-sea mask}
\format{
A climate4R grid object
}
\source{
ERA5 reanalysis data
}
\description{
A binary climate4R grid with the ERA5 land-sea mask (1 = land, 0 = sea).
}
\keyword{data}
