% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utciGrid.R
\name{utciGrid}
\alias{utciGrid}
\title{Universal Thermal Climate Index in climate4R}
\usage{
utciGrid(
  tmax,
  tmin,
  hurs,
  wind,
  radiation,
  mask = NULL,
  parallel = FALSE,
  max.ncores = 16,
  ncores = NULL
)
}
\arguments{
\item{tmax}{A climate4R dataset of daily maximum air temperature.}

\item{tmin}{A climate4R dataset of daily minimum air temperature.}

\item{hurs}{A climate4R dataset of daily relative humidity.}

\item{wind}{A climate4R dataset of daily wind speed.}

\item{radiation}{A climate4R dataset of daily radiation.}

\item{mask}{Optional. A binary climate4R grid (0 and 1) with lat x lon dimensions to mask the output.
UTCI will be computed at gridboxes with mask value 1, while the rest will be set to NA.
If NULL (default), UTCI is calculated at all gridboxes (sea values may be inaccurate).}

\item{parallel}{Logical. Enable parallel processing (default = FALSE).}

\item{max.ncores}{Integer. Maximum number of cores to use (default = 16).}

\item{ncores}{Integer. Specific number of cores to use.}
}
\value{
A climate4R object with the computed index.
}
\description{
Computation of the Universal Thermal Climate Index (UTCI) by ECA&D (European Climate
Assessment & Dataset) directly from climate4R objects.
}
\details{
Air temperature datasets are internally converted to degC, while relative humidity is converted to \%,
wind speed to m s-1 and radiation to MJ m-2.
The Fortran backend \code{calc_UTCI_vector} of the \pkg{UTCIr} package is internally called.
}
\examples{
\dontrun{
require(climate4R.UTCI)

data("ERA5_day_t2mx", package = "climate4R.UTCI")
data("ERA5_day_t2mn", package = "climate4R.UTCI")
data("ERA5_day_hurs", package = "climate4R.UTCI")
data("ERA5_day_sfcwind", package = "climate4R.UTCI")
data("ERA5_day_ssrd", package = "climate4R.UTCI")
data("ERA5_lsm", package = "climate4R.UTCI")       # optional land-sea mask

utci <- utciGrid(tmax = ERA5_day_t2mx, tmin = ERA5_day_t2mn,
                 hurs = ERA5_day_hurs, wind = ERA5_day_sfcwind,
                 radiation = ERA5_day_ssrd, mask = ERA5_lsm)
}

}
\references{
Bröde, P. et al. (2012). Deriving the operational procedure for the Universal Thermal
Climate Index (UTCI). \emph{International Journal of Biometeorology}, 56, 481-494.
\doi{10.1007/s00484-011-0454-1}
}
\author{
climate4R adaptation by C. Rodriguez-Rumayor.
Original \pkg{UTCIr} by ECA-D (\url{https://github.com/ECA-D/UTCIr}).
}
